TRUS-3378 - Validating that property or land trust value is less than total value

+++ app/mapping/ShareAssetMapper.scala
@@ -21,7 +21,8 @@ import mapping.reads.{Asset, ShareNonPortfolioAsset, SharePortfolioAsset}
 import models.WhatKindOfAsset.Shares
 import models.{ShareClass, SharesType, UserAnswers}
 
-class ShareAssetMapper @Inject() extends Mapping[List[SharesType]]{
+class ShareAssetMapper @Inject() extends Mapping[List[SharesType]] {
+
   override def build(userAnswers: UserAnswers): Option[List[SharesType]] = {
 
     val shares: List[Asset] =
@@ -34,10 +35,10 @@ class ShareAssetMapper @Inject() extends Mapping[List[SharesType]]{
       case list =>
         Some(
           list.flatMap {
-            case x : ShareNonPortfolioAsset =>
-              Some(SharesType(x.quantityInTheTrust, x.shareCompanyName, ShareClass.toDES(x.`class`), x.quoted, x.value.toLong))
-            case x : SharePortfolioAsset =>
-              Some(SharesType(x.quantityInTheTrust, x.name, ShareClass.toDES(ShareClass.Other), x.quoted, x.value.toLong))
+            case x: ShareNonPortfolioAsset =>
+              Some(SharesType(x.quantityInTheTrust, x.shareCompanyName, ShareClass.toDES(x.`class`), x.quoted, x.value))
+            case x: SharePortfolioAsset =>
+              Some(SharesType(x.quantityInTheTrust, x.name, ShareClass.toDES(ShareClass.Other), x.quoted, x.value))
             case _ => None
           }
         )
